Add text content rendering support for `InsExtension`
See #170.
