+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html),
 with the exception that 0.x versions can break between minor versions.
 
+## Unreleased
+### Added
+- Add text content rendering support for `InsExtension`
+
 ## [0.15.0] - 2020-05-21
 ### Added
 - Extension for width/height attributes for images, thanks @dohertyfjatl
