+++ commonmark-ext-ins/src/main/java/org/commonmark/ext/ins/InsExtension.java
@@ -2,12 +2,16 @@ package org.commonmark.ext.ins;
 
 import org.commonmark.Extension;
 import org.commonmark.ext.ins.internal.InsDelimiterProcessor;
-import org.commonmark.ext.ins.internal.InsNodeRenderer;
+import org.commonmark.ext.ins.internal.InsHtmlNodeRenderer;
+import org.commonmark.ext.ins.internal.InsTextContentNodeRenderer;
+import org.commonmark.parser.Parser;
+import org.commonmark.renderer.NodeRenderer;
 import org.commonmark.renderer.html.HtmlNodeRendererContext;
 import org.commonmark.renderer.html.HtmlNodeRendererFactory;
-import org.commonmark.parser.Parser;
 import org.commonmark.renderer.html.HtmlRenderer;
-import org.commonmark.renderer.NodeRenderer;
+import org.commonmark.renderer.text.TextContentNodeRendererContext;
+import org.commonmark.renderer.text.TextContentNodeRendererFactory;
+import org.commonmark.renderer.text.TextContentRenderer;
 
 /**
  * Extension for ins using ++
@@ -20,7 +24,9 @@ import org.commonmark.renderer.NodeRenderer;
  * The parsed ins text regions are turned into {@link Ins} nodes.
  * </p>
  */
-public class InsExtension implements Parser.ParserExtension, HtmlRenderer.HtmlRendererExtension {
+public class InsExtension implements Parser.ParserExtension,
+        HtmlRenderer.HtmlRendererExtension,
+        TextContentRenderer.TextContentRendererExtension {
 
     private InsExtension() {
     }
@@ -39,7 +45,17 @@ public class InsExtension implements Parser.ParserExtension, HtmlRenderer.HtmlRe
         rendererBuilder.nodeRendererFactory(new HtmlNodeRendererFactory() {
             @Override
             public NodeRenderer create(HtmlNodeRendererContext context) {
-                return new InsNodeRenderer(context);
+                return new InsHtmlNodeRenderer(context);
+            }
+        });
+    }
+
+    @Override
+    public void extend(TextContentRenderer.Builder rendererBuilder) {
+        rendererBuilder.nodeRendererFactory(new TextContentNodeRendererFactory() {
+            @Override
+            public NodeRenderer create(TextContentNodeRendererContext context) {
+                return new InsTextContentNodeRenderer(context);
             }
         });
     }
